Fix method

+++ lab1/source/dynamic_count.h
@@ -3,6 +3,6 @@
 #include <string.h>
 #include <stdio.h>
 
-typedef int** DynamicTable;
+typedef size_t** DynamicTable;
 
-extern int countNumbers(int n, int k);
+extern size_t countNumbers(size_t n, size_t k);
